refactor: relocate local storage utilities
